adds home to the new router
