+++ src/routes.ts
@@ -24,6 +24,7 @@ import { createChatRouter } from './routes/chat'
 import { createNotificationsRouter } from './routes/notifications'
 import { createSearchRouter } from './routes/search'
 import { createStatusRouter } from './routes/status'
+import { createHomeRouter } from './routes/home'
 
 const limiter = rateLimit({
 	windowMs: 60 * 60 * 1000,
@@ -94,6 +95,7 @@ export const createRouter = (ctx: AppContext) => {
 	router.use(createNotificationsRouter(ctx))
 	router.use(createSearchRouter(ctx))
 	router.use(createStatusRouter(ctx))
+	router.use(createHomeRouter(ctx))
 
 	// OAuth metadata
 	router.get(
@@ -175,65 +177,6 @@ export const createRouter = (ctx: AppContext) => {
 			});
 			await session.destroy();
 			return res.redirect("/");
-		}),
-	);
-
-	// Homepage
-	router.get(
-		"/",
-		handler(async (req, res) => {
-			// If the user is signed in, get an agent which communicates with their server
-			const agent = await getSessionAgent(req, res, ctx);
-
-			// Fetch data stored in mongodb
-			const db = ctx.dbm.db("statusphere");
-			const collection = db.collection("status");
-			const statuses = await collection
-				.find({})
-				.sort({ indexedAt: -1 })
-				.toArray()
-				.then((docs) =>
-					docs.map((doc) => ({
-						uri: doc.uri,
-						authorDid: doc.authorDid,
-						status: doc.status,
-						createdAt: doc.createdAt,
-						indexedAt: doc.indexedAt,
-					})),
-				);
-
-			// Map user DIDs to their domain-name handles
-			const didHandleMap = await ctx.resolver.resolveDidsToHandles(
-				statuses.map((s) => s.authorDid),
-			);
-
-			if (!agent) {
-				// Serve the logged-out view
-				return res.type("html").send(page(login({})));
-			}
-
-			// Fetch additional information about the logged-in user
-			const { data: profileRecord } = await agent.com.atproto.repo.getRecord({
-				repo: agent.assertDid,
-				collection: "app.bsky.actor.profile",
-				rkey: "self",
-			});
-			const profile =
-				Profile.isRecord(profileRecord.value) &&
-				Profile.validateRecord(profileRecord.value).success
-					? profileRecord.value
-					: {};
-
-			// Serve the logged-in view
-			return res.type("html").send(
-				page(
-					home({
-						statuses,
-						didHandleMap,
-						profile,
-					}),
-				),
-			);
 		}),
 	);
